Get passwords before environment file is saved
